Refactor settings to individual files
