+++ packages/shared/components/SettingsMenu.svelte
@@ -1,8 +1,6 @@
 <script lang="typescript">
     import { Icon, Text } from 'shared/components'
-    import { Locale } from 'shared/lib/typings/i18n'
-
-    export let locale: Locale
+    import { localize } from 'shared/lib/i18n'
 
     export let settings
     export let activeSettings = []
@@ -25,7 +23,9 @@
             class={`group flex flex-row justify-start items-center hover:bg-blue-50 dark:hover:bg-gray-800 dark:hover:bg-opacity-20 py-2 rounded-lg w-3/4 text-left ${Object.values(activeSettings).includes(setting) ? '' : 'opacity-20 pointer-events-none'}`}
             disabled={!Object.values(activeSettings).includes(setting)}>
             <Icon icon={icons[setting]} classes="text-gray-500 ml-1 mr-3 group-hover:text-blue-500" />
-            <Text type="p" secondary classes="group-hover:text-blue-500">{locale(`views.settings.${setting}.title`)}</Text>
+            <Text type="p" secondary classes="group-hover:text-blue-500">
+                {localize(`views.settings.${setting}.title`)}
+            </Text>
         </button>
     {/each}
 </div>
