+++ packages/shared/components/SettingsNavigator.svelte
@@ -1,8 +1,6 @@
 <script lang="typescript">
     import { Icon, Text } from 'shared/components'
-    import { Locale } from 'shared/lib/typings/i18n'
-
-    export let locale: Locale
+    import { localize } from 'shared/lib/i18n'
 
     export let settings
     export let routes
@@ -28,7 +26,7 @@
                         icon="small-chevron-right"
                         classes="text-blue-500 absolute left-1 top-0.4 text-xl" />
                 {/if}
-                <Text type="p" classes="mb-1">{locale(`views.settings.${setting}.title`)}</Text>
+                <Text type="p" classes="mb-1">{localize(`views.settings.${setting}.title`)}</Text>
             </button>
             {#if route === setting}
                 {#each Object.values(settings[route]) as setting, i}
@@ -36,7 +34,9 @@
                         on:click={() => onSettingClick(setting)}
                         class="group flex flex-row justify-start items-center hover:bg-blue-50 dark:hover:bg-gray-800 dark:hover:bg-opacity-20 py-2 rounded-lg w-full ml-6 text-left {Object.values(settings[route]).length - 1 === i && 'mb-4'}">
                         <Icon icon={icons[setting]} classes="text-gray-500 ml-1 mr-3 group-hover:text-blue-500" />
-                        <Text type="p" classes="group-hover:text-blue-500">{locale(`views.settings.${setting}.title`)}</Text>
+                        <Text type="p" classes="group-hover:text-blue-500">
+                            {localize(`views.settings.${setting}.title`)}
+                        </Text>
                     </button>
                 {/each}
             {/if}
