+++ packages/shared/lib/typings/routes.ts
@@ -77,8 +77,6 @@ export enum SettingsRoutesNoProfile {
 }
 
 export enum GeneralSettings {
-    // TODO: Implement and enable
-    // Profile = 'profile',
     Theme = 'theme',
     Language = 'language',
     Currency = 'currency',
@@ -102,7 +100,6 @@ export enum SecuritySettings {
 
 export enum AdvancedSettings {
     NetworkConfiguration = 'networkConfiguration',
-    // DeveloperMode = 'developerMode',
     DeepLinks = 'deepLinks',
     BalanceFinder = 'balanceFinder',
     HiddenAccounts = 'hiddenAccounts',
@@ -112,16 +109,21 @@ export enum AdvancedSettings {
 }
 
 export enum AdvancedSettingsNoProfile {
-    // DeveloperMode = 'developerMode',
     DeepLinks = 'deepLinks',
     ErrorLog = 'errorLog',
     Diagnostics = 'diagnostics',
 }
 
 export enum HelpAndInfo {
-    // Troubleshoot = 'troubleshoot',
     Documentation = 'documentation',
     FAQ = 'faq',
     Discord = 'discord',
     ReportAnIssue = 'reportAnIssue',
+}
+
+export enum ExternalRoute {
+    Documentation = 'https://chrysalis.docs.iota.org/',
+    Discord = 'https://discord.iota.org',
+    FAQ = 'https://firefly.iota.org/faq',
+    IssueReport = 'https://github.com/iotaledger/firefly/issues',
 }
