+++ packages/shared/routes/dashboard/settings/views/SettingsHome.svelte
@@ -1,6 +1,7 @@
 <script lang="typescript">
     import { SettingsMenu, Text } from 'shared/components'
     import { loggedIn } from 'shared/lib/app'
+    import { localize } from 'shared/lib/i18n'
     import { isLedgerProfile, isSoftwareProfile } from 'shared/lib/profile'
     import { settingsChildRoute, settingsRoute } from 'shared/lib/router'
     import { SettingsIcons } from 'shared/lib/typings/icons'
@@ -13,11 +14,6 @@
         SecuritySettings,
         SettingsRoutes,
     } from 'shared/lib/typings/routes'
-    import { Locale } from 'shared/lib/typings/i18n'
-
-    export let locale: Locale
-
-    export let mobile
 
     const securitySettings = Object.assign({}, SecuritySettings)
     const advancedSettings = Object.assign({}, AdvancedSettings)
@@ -30,66 +26,56 @@
     $: if (!$isLedgerProfile) {
         delete advancedSettings.MigrateLedgerIndex
     }
+
+    function onSettingClick(
+        route: SettingsRoutes,
+        childRoute:
+            | SecuritySettings
+            | AdvancedSettings
+            | GeneralSettings
+            | GeneralSettingsNoProfile
+            | AdvancedSettingsNoProfile
+            | HelpAndInfo
+    ) {
+        settingsRoute.set(route)
+        settingsChildRoute.set(childRoute)
+    }
 </script>
 
-{#if mobile}
-    <div>foo</div>
-{:else}
-    <div class="h-full w-full flex flex-col">
-        <Text type="h2" classes="mb-14">{locale('views.settings.settings')}</Text>
-        <div class="flex flex-row items-start space-x-10">
-            <SettingsMenu
-                icon="settings"
-                iconColor="bg-blue-500"
-                icons={SettingsIcons}
-                settings={GeneralSettings}
-                activeSettings={$loggedIn ? GeneralSettings : GeneralSettingsNoProfile}
-                title={locale('views.settings.generalSettings.title')}
-                description={locale('views.settings.generalSettings.description')}
-                onClick={(setting) => {
-                    settingsRoute.set(SettingsRoutes.GeneralSettings)
-                    settingsChildRoute.set(setting)
-                }}
-                {locale} />
-            <SettingsMenu
-                icon="security"
-                iconColor="bg-yellow-500"
-                icons={SettingsIcons}
-                settings={securitySettings}
-                activeSettings={$loggedIn ? SecuritySettings : undefined}
-                title={locale('views.settings.security.title')}
-                description={locale('views.settings.security.description')}
-                onClick={(setting) => {
-                    settingsRoute.set(SettingsRoutes.Security)
-                    settingsChildRoute.set(setting)
-                }}
-                {locale} />
-            <SettingsMenu
-                icon="tools"
-                iconColor="bg-green-600"
-                icons={SettingsIcons}
-                settings={advancedSettings}
-                activeSettings={$loggedIn ? advancedSettings : AdvancedSettingsNoProfile}
-                title={locale('views.settings.advancedSettings.title')}
-                description={locale('views.settings.advancedSettings.description')}
-                onClick={(setting) => {
-                    settingsRoute.set(SettingsRoutes.AdvancedSettings)
-                    settingsChildRoute.set(setting)
-                }}
-                {locale} />
-            <SettingsMenu
-                icon="info"
-                iconColor="bg-purple-500"
-                icons={SettingsIcons}
-                settings={HelpAndInfo}
-                activeSettings={HelpAndInfo}
-                title={locale('views.settings.helpAndInfo.title')}
-                description={locale('views.settings.helpAndInfo.description')}
-                onClick={(setting) => {
-                    settingsRoute.set(SettingsRoutes.HelpAndInfo)
-                    settingsChildRoute.set(setting)
-                }}
-                {locale} />
-        </div>
+<div class="flex flex-col flex-initial h-full w-full">
+    <Text type="h2" classes="mb-14">{localize('views.settings.settings')}</Text>
+    <div class="flex items-start flex-row space-x-10">
+        <SettingsMenu
+            icons={SettingsIcons}
+            settings={GeneralSettings}
+            activeSettings={$loggedIn ? GeneralSettings : GeneralSettingsNoProfile}
+            title={localize('views.settings.generalSettings.title')}
+            description=""
+            onClick={(setting) => onSettingClick(SettingsRoutes.GeneralSettings, setting)}
+            {localize} />
+        <SettingsMenu
+            icons={SettingsIcons}
+            settings={securitySettings}
+            activeSettings={$loggedIn ? SecuritySettings : undefined}
+            title={localize('views.settings.security.title')}
+            description=""
+            onClick={(setting) => onSettingClick(SettingsRoutes.Security, setting)}
+            {localize} />
+        <SettingsMenu
+            icons={SettingsIcons}
+            settings={advancedSettings}
+            activeSettings={$loggedIn ? advancedSettings : AdvancedSettingsNoProfile}
+            title={localize('views.settings.advancedSettings.title')}
+            description=""
+            onClick={(setting) => onSettingClick(SettingsRoutes.AdvancedSettings, setting)}
+            {localize} />
+        <SettingsMenu
+            icons={SettingsIcons}
+            settings={HelpAndInfo}
+            activeSettings={HelpAndInfo}
+            title={localize('views.settings.helpAndInfo.title')}
+            description=""
+            onClick={(setting) => onSettingClick(SettingsRoutes.HelpAndInfo, setting)}
+            {localize} />
     </div>
-{/if}
+</div>
