+++ packages/shared/routes/dashboard/settings/views/advanced/DeveloperMode.svelte
@@ -1,0 +1,14 @@
+<script lang="typescript">
+    import { Checkbox, Text } from 'shared/components'
+    import { localize } from 'shared/lib/i18n'
+    import { activeProfile, updateProfile } from 'shared/lib/profile'
+    import { get } from 'svelte/store'
+
+    let isDeveloperProfile = get(activeProfile)?.isDeveloperProfile
+
+    $: updateProfile('isDeveloperProfile', isDeveloperProfile)
+</script>
+
+<Text type="h4" classes="mb-3">{localize('views.settings.developerMode.title')}</Text>
+<Text type="p" secondary classes="mb-5">{localize('views.settings.developerMode.description')}</Text>
+<Checkbox label={localize('actions.enableDeveloperMode')} bind:checked={isDeveloperProfile} />
