+++ packages/shared/routes/dashboard/settings/views/advanced/index.js
@@ -1,0 +1,8 @@
+export { default as BalanceFinder } from './BalanceFinder.svelte'
+export { default as DeepLinks } from './DeepLinks.svelte'
+export { default as DeveloperMode } from './DeveloperMode.svelte'
+export { default as Diagnostics } from './Diagnostics.svelte'
+export { default as ErrorLog } from './ErrorLog.svelte'
+export { default as HiddenAccounts } from './HiddenAccounts.svelte'
+export { default as MigrateLedgerIndex } from './MigrateLedgerIndex.svelte'
+export { default as NetworkConfiguration } from './NetworkConfiguration.svelte'
