+++ packages/shared/routes/dashboard/settings/views/general/index.js
@@ -1,0 +1,5 @@
+export { default as Currency } from './Currency.svelte'
+export { default as Language } from './Language.svelte'
+export { default as NetworkStatus } from './NetworkStatus.svelte'
+export { default as Notifications } from './Notifications.svelte'
+export { default as Theme } from './Theme.svelte'
