+++ packages/shared/routes/dashboard/settings/views/help/index.js
@@ -1,0 +1,4 @@
+export { default as Discord } from './Discord.svelte'
+export { default as Documentation } from './Documentation.svelte'
+export { default as FAQ } from './FAQ.svelte'
+export { default as ReportAnIssue } from './ReportAnIssue.svelte'
