+++ packages/shared/routes/dashboard/settings/views/index.js
@@ -1,6 +1,6 @@
+export { default as Advanced } from './advanced/Advanced.svelte'
+export { default as General } from './general/General.svelte'
+export { default as Help } from './help/Help.svelte'
+export { default as Security } from './security/Security.svelte'
 export { default as SettingsHome } from './SettingsHome.svelte'
 export { default as SettingsViewer } from './SettingsViewer.svelte'
-export { default as General } from './General.svelte'
-export { default as Security } from './Security.svelte'
-export { default as Advanced } from './Advanced.svelte'
-export { default as Help } from './Help.svelte'
