+++ packages/shared/routes/dashboard/settings/views/security/index.js
@@ -1,0 +1,5 @@
+export { default as AppLock } from './AppLock.svelte'
+export { default as ChangePassword } from './ChangePassword.svelte'
+export { default as ChangePincode } from './ChangePincode.svelte'
+export { default as DeleteProfile } from './DeleteProfile.svelte'
+export { default as ExportStronghold } from './ExportStronghold.svelte'
